refactor: move find-free port so _plugins don't depend on _server
