+++ marimo/_server/start.py
@@ -26,7 +26,6 @@ from marimo._server.registry import LIFESPAN_REGISTRY
 from marimo._server.session_manager import SessionManager
 from marimo._server.tokens import AuthToken
 from marimo._server.utils import (
-    find_free_port,
     initialize_asyncio,
     initialize_fd_limit,
 )
@@ -34,6 +33,7 @@ from marimo._server.uvicorn_utils import initialize_signals
 from marimo._session.model import SessionMode
 from marimo._tracer import LOGGER
 from marimo._utils.lifespans import Lifespans
+from marimo._utils.net import find_free_port
 from marimo._utils.paths import marimo_package_path
 
 DEFAULT_PORT = 2718
