+++ tests/_cli/test_cli_convert.py
@@ -13,7 +13,7 @@ from typing import Any
 
 import pytest
 
-from marimo._server.utils import find_free_port
+from marimo._utils.net import find_free_port
 from marimo._utils.platform import is_windows
 from tests.mocks import snapshotter
